refactor(wasm): extract shared Go/wasm boot sequence into boot.cjs
The runtime instantiation sequence (inject fs/process/crypto globals, load
wasm_exec.js, instantiate dbc.wasm, run, and verify the dbc* API registered)
was duplicated byte-for-byte between index.cjs (ensureRuntime) and worker.cjs.
Any change to the boot protocol had to land in both, and a divergence would be
a silent init bug caught only by whichever path was exercised.

Extract it into a single bootRuntime() in boot.cjs. Each caller keeps its own
error-reporting policy: the in-process path namespaces the rejection with the
dbc-wasm: prefix, the worker posts a {type:"fatal"} message. Add boot.cjs to
the published files list in the build.js package.json generator so it ships in
the tarball.

Smoke tests pass for the in-process, worker, and worker-init-failure paths.
